% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deletetuplycates.R
\name{deleteduplicates}
\alias{deleteduplicates}
\title{A function that checks for duplicates. If there are, it deletes the duplicates from a variable.}
\usage{
deleteduplicates(nametable, variable)
}
\arguments{
\item{nametable}{The data table}

\item{variable}{The table variable}
}
\value{
The function returns the table with the variable without duplicates
}
\description{
A function that checks for duplicates. If there are, it deletes the duplicates from a variable.
}
