% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Renamh.R
\name{renameh}
\alias{renameh}
\title{Rename a variable in a data table}
\usage{
renameh(nametable, variable, newname)
}
\arguments{
\item{nametable}{The data table}

\item{variable}{The variable to be renamed}

\item{newname}{The new variable name}
}
\value{
A data table with the variable renamed
}
\description{
This function renames a variable in a data table.
}
\examples{
# Rename the variable 'old_name' to 'new_name' in the data table 'my_data'
my_data <- renameh(my_data, "old_name", "new_name")

# If the variable does not exist, a warning is issued
renameh(my_data, "nonexistent_variable", "new_name")

}
